Moves the user assets to the user folder inside components.
