+++ app/routes/index.ts
@@ -3,7 +3,7 @@ import { RouteDefinition } from 'angular2/router';
 
 // Components
 import { HomeComponent } from '../components/home/component';
-import { UserComponent } from '../components/user.component';
+import { UserComponent } from '../components/user/component';
 
 export const RouteDefinitions: RouteDefinition[] = [
   {
